Improvements and Android Q support
