+++ android/gradle.properties
@@ -1,4 +1,4 @@
-ReactNativeCameraRoll_compileSdkVersion=28
-ReactNativeCameraRoll_buildToolsVersion=28.0.3
+ReactNativeCameraRoll_compileSdkVersion=29
+ReactNativeCameraRoll_buildToolsVersion=29.0.2
 ReactNativeCameraRoll_targetSdkVersion=27
 ReactNativeCameraRoll_minSdkVersion=16
