+++ android/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Dec 15 22:58:57 CET 2020
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.1.1-all.zip
